+++ test/support.js
@@ -166,19 +166,6 @@ var Support = {
     return envDialect;
   },
 
-  dialectIsMySQL: function(strict) {
-    var envDialect = process.env.DIALECT || 'mysql';
-    if (strict === undefined) {
-      strict = false;
-    }
-
-    if (strict) {
-      return envDialect === 'mysql';
-    } else {
-      return ['mysql', 'mariadb'].indexOf(envDialect) !== -1;
-    }
-  },
-
   getTestDialectTeaser: function(moduleName) {
     var dialect = this.getTestDialect();
 
@@ -210,10 +197,6 @@ var Support = {
 
   expectsql: function(query, expectations) {
     var expectation = expectations[Support.sequelize.dialect.name];
-
-    if (!expectation && Support.sequelize.dialect.name === 'mariadb') {
-      expectation = expectations.mysql;
-    }
 
     if (!expectation) {
       expectation = expectations['default']